requirements.txt
streamlit>=1.34.0
tensorflow>=2.14.0
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
Pillow>=10.0.0
matplotlib>=3.7.0
